Practice_Organized/08_LangChain_Mastery/LangChain_Practice/01_Models/requirements.txt
# LangChain Core
langchain==1.1.0

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini Integration
langchain-google-genai==3.2.0
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub == 0.36.0

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
